+++ modules/features/podcasts/src/main/java/au/com/shiftyjelly/pocketcasts/podcasts/view/folders/FolderCreateFragment.kt
@@ -1,10 +1,15 @@
 package au.com.shiftyjelly.pocketcasts.podcasts.view.folders
 
-import android.app.Dialog
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.ViewGroup
+import androidx.compose.foundation.layout.WindowInsets
+import androidx.compose.foundation.layout.WindowInsetsSides
+import androidx.compose.foundation.layout.only
+import androidx.compose.foundation.layout.safeDrawing
+import androidx.compose.foundation.layout.windowInsetsPadding
 import androidx.compose.material.MaterialTheme
+import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.toArgb
 import androidx.core.os.bundleOf
 import androidx.fragment.app.activityViewModels
@@ -66,7 +71,11 @@ class FolderCreateFragment : BaseDialogFragment() {
 
             navHostController = rememberNavController()
             val navController = navHostController ?: return@AppThemeWithBackground
-            NavHost(navController = navController, startDestination = NavRoutes.PODCASTS) {
+            NavHost(
+                navController = navController,
+                startDestination = NavRoutes.PODCASTS,
+                modifier = Modifier.windowInsetsPadding(WindowInsets.safeDrawing.only(WindowInsetsSides.Horizontal)),
+            ) {
                 composable(NavRoutes.PODCASTS) {
                     FolderEditPodcastsPage(
                         onCloseClick = { dismiss() },