+++ CHANGELOG.md
@@ -11,6 +11,8 @@
         ([#4562](https://github.com/Automattic/pocket-casts-android/pull/4562))
     *   Fix done button overlapping with navigation bars
         ([#4553](https://github.com/Automattic/pocket-casts-android/pull/4553))
+    *   Fix folder flow buttons overlapping with navigation bars
+        ([#4597](https://github.com/Automattic/pocket-casts-android/pull/4597))
     *   Fix shared list under the mini player
         ([#4560](https://github.com/Automattic/pocket-casts-android/pull/4560))
     *   Fix transcript button incorrectly disabled